samples/metadetect/requirements.txt
opencv-python-headless
numpy>=1.19.5
Pillow>=8.0.0
pyzmq>=26.3.0
ultralytics>=8.0.0
